优化支持文件分片、断点续传、文件秒传代码

+++ ruoyi-system/src/main/java/com/ruoyi/system/service/file/impl/BackFilelistServiceImpl.java
@@ -1,10 +1,12 @@
 package com.ruoyi.system.service.file.impl;
 
-import com.ruoyi.common.utils.file.FileUtils;
 import com.ruoyi.system.domain.BackFilelist;
 import com.ruoyi.system.mapper.BackFilelistMapper;
 import com.ruoyi.system.service.file.IBackFilelistService;
 import org.apache.commons.compress.utils.Lists;
+import org.apache.commons.io.FileUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -22,6 +24,8 @@ import java.util.List;
 @Service
 public class BackFilelistServiceImpl implements IBackFilelistService
 {
+    private static final Logger log = LoggerFactory.getLogger(BackFilelistServiceImpl.class);
+
     @Autowired
     private BackFilelistMapper backFilelistMapper;
 
@@ -31,7 +35,7 @@ public class BackFilelistServiceImpl implements IBackFilelistService
     private final static String folderPath = "/file";
     /**
      * 查询已上传文件列表
-     * 
+     *
      * @param id 已上传文件列表ID
      * @return 已上传文件列表
      */
@@ -43,7 +47,7 @@ public class BackFilelistServiceImpl implements IBackFilelistService
 
     /**
      * 查询已上传文件列表列表
-     * 
+     *
      * @param backFilelist 已上传文件列表
      * @return 已上传文件列表
      */
@@ -55,7 +59,7 @@ public class BackFilelistServiceImpl implements IBackFilelistService
 
     /**
      * 新增已上传文件列表
-     * 
+     *
      * @param backFilelist 已上传文件列表
      * @return 结果
      */
@@ -67,7 +71,7 @@ public class BackFilelistServiceImpl implements IBackFilelistService
 
     /**
      * 修改已上传文件列表
-     * 
+     *
      * @param backFilelist 已上传文件列表
      * @return 结果
      */
@@ -79,26 +83,40 @@ public class BackFilelistServiceImpl implements IBackFilelistService
 
     /**
      * 批量删除已上传文件列表
-     * 
+     *
      * @param ids 需要删除的已上传文件列表ID
      * @return 结果
      */
     @Override
     public int deleteBackFilelistByIds(Long[] ids)
     {
-        ArrayList<Boolean> objects = Lists.newArrayList();
+        ArrayList<Boolean> dels = Lists.newArrayList();
         //先删除文件再删除对应的表记录
         for (Long id : ids) {
             BackFilelist backFilelist = backFilelistMapper.selectBackFilelistById(id);
             String fileFolder = filePath + folderPath + "/" + backFilelist.getIdentifier();
-            FileUtils.deleteFolder(new File(fileFolder));
+            String filePathTrue=fileFolder+"/"+backFilelist.getFilename();
+            boolean b=false;
+            try {
+                b = Files.deleteIfExists(Paths.get(filePathTrue));
+                dels.add(b);
+              /*  if(b){
+                    FileUtils.deleteDirectory(new File(fileFolder));
+                }*/
+            } catch (IOException e) {
+                log.info("删除文件: {}报错，信息为{}", filePathTrue,e.getMessage());
+            }
+        }
+        if(dels.contains(false)){
+            //删除失败
+            return 0;
         }
         return backFilelistMapper.deleteBackFilelistByIds(ids);
     }
 
     /**
      * 删除已上传文件列表信息
-     * 
+     *
      * @param id 已上传文件列表ID
      * @return 结果
      */
